move to textual inversion mixin
